incorporate review feedback: remove tsUtils  and update addEmailToForm
